Add view methods to IERC721M (#33)
* Add view methods to IERC721M

* fix timestamp size

+++ test/erc721m.test.ts
@@ -702,7 +702,7 @@ describe('ERC721M', function () {
 
       const timestamp = stageStart + 500;
       const digestFromJs = ethers.utils.solidityKeccak256(
-        ['address', 'address', 'uint32', 'address', 'uint256'],
+        ['address', 'address', 'uint32', 'address', 'uint64'],
         [contract.address, minter.address, 1, cosigner.address, timestamp],
       );
       const sig = await cosigner.signMessage(
@@ -741,7 +741,7 @@ describe('ERC721M', function () {
 
       const timestamp = Math.floor(new Date().getTime() / 1000);
       const digestFromJs = ethers.utils.solidityKeccak256(
-        ['address', 'address', 'uint32', 'address', 'uint256'],
+        ['address', 'address', 'uint32', 'address', 'uint64'],
         [contract.address, minter.address, 1, cosigner.address, timestamp],
       );
       const sig = await cosigner.signMessage(
@@ -836,7 +836,7 @@ describe('ERC721M', function () {
 
       const earlyTimestamp = stageStart - 1;
       let digestFromJs = ethers.utils.solidityKeccak256(
-        ['address', 'address', 'uint32', 'address', 'uint256'],
+        ['address', 'address', 'uint32', 'address', 'uint64'],
         [contract.address, minter.address, 1, cosigner.address, earlyTimestamp],
       );
       let sig = await cosigner.signMessage(ethers.utils.arrayify(digestFromJs));
@@ -855,7 +855,7 @@ describe('ERC721M', function () {
 
       const lateTimestamp = stageStart + 1001;
       digestFromJs = ethers.utils.solidityKeccak256(
-        ['address', 'address', 'uint32', 'address', 'uint256'],
+        ['address', 'address', 'uint32', 'address', 'uint64'],
         [contract.address, minter.address, 1, cosigner.address, lateTimestamp],
       );
       sig = await cosigner.signMessage(ethers.utils.arrayify(digestFromJs));
@@ -894,7 +894,7 @@ describe('ERC721M', function () {
 
       const timestamp = stageStart;
       const digestFromJs = ethers.utils.solidityKeccak256(
-        ['address', 'address', 'uint32', 'address', 'uint256'],
+        ['address', 'address', 'uint32', 'address', 'uint64'],
         [contract.address, minter.address, 1, cosigner.address, timestamp],
       );
       const sig = await cosigner.signMessage(
@@ -1528,7 +1528,7 @@ describe('ERC721M', function () {
 
       const timestamp = Math.floor(new Date().getTime() / 1000);
       const digestFromJs = ethers.utils.solidityKeccak256(
-        ['address', 'address', 'uint32', 'address', 'uint256'],
+        ['address', 'address', 'uint32', 'address', 'uint64'],
         [erc721M.address, minter.address, 1, cosigner.address, timestamp],
       );
       const sig = await cosigner.signMessage(
